requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
alembic-autogenerate-enums==0.1.2
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.30.0
attrs==24.2.0
click==8.1.7
docker==7.1.0
dependency-injector==4.42.0
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
httpx==0.27.2
idna==3.10
loguru==0.7.2
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pytest==8.3.3
python-decouple==3.8
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
yarl==1.11.1
